Refactored dashboard, created new directories for components and renamed to include .jsx
