+++ src/components/Exercise/exercise-results.jsx
@@ -1,11 +1,11 @@
 import React, {Component} from 'react';
-import {clearAuth} from '../actions/auth';
-import {clearAuthToken} from '../local-storage';
-import NutritionSearchPage from './nutrition-search-page';
+import {clearAuth} from '../../actions/auth';
+import {clearAuthToken} from '../../local-storage';
+import NutritionSearchPage from '../nutrition-search-page';
 import ExerciseSearchPage from './exercise-search-page';
 import {connect} from 'react-redux';
-import {addExercise} from '../actions/protected-exercise-data';
-import requiresLogin from './requires-login';
+import {addExercise} from '../../actions/protected-exercise-data';
+import requiresLogin from '../Login/requires-login';
 import '../index.scss';
 
 export class ExerciseResults extends Component {
