+++ src/components/Nutrition/nutrition-results.jsx
@@ -1,12 +1,12 @@
 import React, {Component} from 'react';
-import {clearAuth} from '../actions/auth';
-import {clearAuthToken} from '../local-storage';
+import {clearAuth} from '../../actions/auth';
+import {clearAuthToken} from '../../local-storage';
 import '../index.scss';
 import NutritionSearchPage from './nutrition-search-page';
 import ExerciseSearchPage from './exercise-search-page';
 import {connect} from 'react-redux';
-import {addProtectedData} from '../actions/protected-data';
-import requiresLogin from './requires-login';
+import {addProtectedData} from '../../actions/protected-data';
+import requiresLogin from '../Login/requires-login';
 
 export class NutritionResults extends Component {
   constructor(props) {
