+++ src/components/Registration/registration-page.jsx
