--- src/util.js
@@ -1,4 +1,0 @@
-export const updateObject = (oldObject, updatedProperties) => ({
-  ...oldObject,
-  ...updatedProperties
-});
